+++ Source/Editor/Foundation/StandardFileTypes.cpp
@@ -109,7 +109,7 @@ void Foundation_StandardFileTypes(Context* context, Project* project)
 
     project->AddAnalyzeFileCallback([](ResourceFileDescriptor& desc, const AnalyzeFileContext& ctx)
     {
-        if (desc.HasExtension({".mdl", ".obj" }))
+        if (desc.HasExtension({".mdl" }))
         {
             desc.AddObjectType<Model>();
         }
