+++ SourceAssets/ObjTests/FlipYZ.obj
@@ -1,0 +1,44 @@
+mtllib FlipYZ.mtl
+
+#
+# object TestPiramid
+#
+
+v  -5.0000 -5.0000 5.0000
+v  0.0000 -5.0000 -10.0000
+v  5.0000 -5.0000 5.0000
+v  0.0000 5.0000 -0.0000
+# 4 vertices
+
+vn -0.7441 -0.3398 0.5752
+vn 0.0000 -0.2672 -0.9636
+vn 0.7441 -0.3398 0.5752
+vn 0.0000 0.9999 -0.0160
+# 4 vertex normals
+
+vt 1.0000 0.0000 0.0000
+vt 0.5000 1.0000 0.0000
+vt 0.0000 0.0000 0.0000
+vt 0.5000 1.0000 0.3333
+vt 0.0000 0.0000 1.0000
+vt 1.0000 0.0000 0.5000
+vt 0.3333 1.0000 0.5000
+vt 0.0000 0.0000 0.5000
+vt 0.6667 1.0000 0.5000
+# 9 texture coords
+
+o TestPiramid
+g TestPiramid
+usemtl Common
+s 1
+f 1/1/1 2/2/2 3/3/3 
+s 2
+f 1/3/1 3/1/3 4/4/4 
+usemtl Red
+s 4
+f 3/5/3 2/6/2 4/7/4 
+usemtl Common
+s 8
+f 2/8/2 1/1/1 4/9/4 
+# 0 polygons - 4 triangles
+
