+++ SourceAssets/ObjTests/NoNormals.obj
@@ -1,0 +1,38 @@
+mtllib NoNormals.mtl
+
+#
+# object TestPiramid
+#
+
+v  -5.0000 -5.0000 -5.0000
+v  0.0000 10.0000 -5.0000
+v  5.0000 -5.0000 -5.0000
+v  0.0000 0.0000 5.0000
+# 4 vertices
+
+vt 1.0000 0.0000 0.0000
+vt 0.5000 1.0000 0.0000
+vt 0.0000 0.0000 0.0000
+vt 0.5000 1.0000 0.3333
+vt 0.0000 0.0000 1.0000
+vt 1.0000 0.0000 0.5000
+vt 0.3333 1.0000 0.5000
+vt 0.0000 0.0000 0.5000
+vt 0.6667 1.0000 0.5000
+# 9 texture coords
+
+o TestPiramid
+g TestPiramid
+usemtl Common
+s 1
+f 1/1 2/2 3/3 
+s 2
+f 1/3 3/1 4/4 
+usemtl Red
+s 4
+f 3/5 2/6 4/7 
+usemtl Common
+s 8
+f 2/8 1/1 4/9 
+# 0 polygons - 4 triangles
+
