+++ SourceAssets/ObjTests/NoUV.obj
@@ -1,0 +1,33 @@
+mtllib NoUV.mtl
+
+#
+# object TestPiramid
+#
+
+v  -5.0000 -5.0000 -5.0000
+v  0.0000 10.0000 -5.0000
+v  5.0000 -5.0000 -5.0000
+v  0.0000 0.0000 5.0000
+# 4 vertices
+
+vn -0.7441 -0.5752 -0.3398
+vn 0.0000 0.9636 -0.2672
+vn 0.7441 -0.5752 -0.3398
+vn 0.0000 0.0160 0.9999
+# 4 vertex normals
+
+o TestPiramid
+g TestPiramid
+usemtl Common
+s 1
+f 1//1 2//2 3//3 
+s 2
+f 1//1 3//3 4//4 
+usemtl Red
+s 4
+f 3//3 2//2 4//4 
+usemtl Common
+s 8
+f 2//2 1//1 4//4 
+# 0 polygons - 4 triangles
+
